Refactor: Modal 리팩토링
