+++ src/pages/Home.tsx
@@ -6,9 +6,9 @@ import InputWithCalendar from "@/components/common/InputWithCalendar"; //FIXME: 
 import FactorSetup from "@/components/Backtesting/FactorSetup";
 import TechnicalAnalysisStrategySetup from "@/components/Backtesting/TechnicalAnalysisStrategySetup";
 import Modal from "@/components/Modal/Modal";
-import { useState } from "react";
 import ModalInner from "@/components/Modal/ModalInner";
 import MyStrategyList from "@/components/myStrategyFAB/MyStrategyList";
+import useModal from "@/hooks/useModal";
 
 //FIXME: 화면 확인용
 const OPTIONS = [
@@ -18,11 +18,8 @@ const OPTIONS = [
 ];
 
 const Home = () => {
-  const [isOpen, setIsOpen] = useState(true);
-
-  const handleModalClose = () => {
-    setIsOpen(false);
-  };
+  const { modalRef, isOpen, handleModalClose, handleOverlayClick } =
+    useModal(true);
 
   return (
     <HomeStyle>
@@ -34,7 +31,11 @@ const Home = () => {
       <InputWithCalendar />
       <FactorSetup />
       <TechnicalAnalysisStrategySetup />
-      <Modal isOpen={isOpen} onClose={handleModalClose}>
+      <Modal
+        modalRef={modalRef}
+        isOpen={isOpen}
+        handleOverlayClick={handleOverlayClick}
+      >
         <ModalInner onClose={handleModalClose}>모달창</ModalInner>
       </Modal>
       <MyStrategyList />
